[SHELL32] Separate CCopyAsPathMenu implementation from CCopyMoveToMenu
